refactor: create facade layer for business logic seperate from service layer
