+++ backend/src/main/java/com/chillin/hearting/api/controller/HeartController.java
@@ -4,8 +4,7 @@ import com.chillin.hearting.api.data.Data;
 import com.chillin.hearting.api.data.HeartData;
 import com.chillin.hearting.api.data.HeartListData;
 import com.chillin.hearting.api.response.ResponseDTO;
-import com.chillin.hearting.api.service.HeartService;
-import com.chillin.hearting.api.service.UserHeartService;
+import com.chillin.hearting.api.service.facade.HeartFacade;
 import com.chillin.hearting.db.domain.User;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -22,8 +21,8 @@ import java.util.List;
 @RequiredArgsConstructor
 public class HeartController {
 
-    private final HeartService heartService;
-    private final UserHeartService userHeartService;
+    private final HeartFacade heartFacade;
+
     private static final String MESSAGE_SUCCESS = "success";
     private static final String FIND_ALLHEARTS_SUCCESS = "도감용 하트 리스트 조회에 성공했습니다.";
     private static final String FIND_MSGHEARTS_SUCCESS = "메시지용 하트 리스트 조회에 성공했습니다.";
@@ -34,7 +33,7 @@ public class HeartController {
     @GetMapping("")
     public ResponseEntity<ResponseDTO> findAllHearts(HttpServletRequest httpServletRequest) {
         User user = (User) httpServletRequest.getAttribute("user");
-        Data data = heartService.findAllHearts(user);
+        Data data = heartFacade.findAllHearts(user);
         ResponseDTO responseDTO = ResponseDTO.builder().status(MESSAGE_SUCCESS).message(FIND_ALLHEARTS_SUCCESS).data(data).build();
         return new ResponseEntity<>(responseDTO, HttpStatus.OK);
     }
@@ -42,7 +41,7 @@ public class HeartController {
     @GetMapping("/user-hearts")
     public ResponseEntity<ResponseDTO> findUserHearts(HttpServletRequest httpServletRequest) {
         User user = (User) httpServletRequest.getAttribute("user");
-        List<HeartData> messageHearts = heartService.findMessageHearts(user);
+        List<HeartData> messageHearts = heartFacade.findMessageHearts(user);
         ResponseDTO responseDTO = ResponseDTO.builder().status(MESSAGE_SUCCESS).message(FIND_MSGHEARTS_SUCCESS).data(HeartListData.builder().heartList(messageHearts).build()).build();
         return new ResponseEntity<>(responseDTO, HttpStatus.OK);
     }
@@ -50,7 +49,7 @@ public class HeartController {
     @GetMapping("/{heartId}")
     public ResponseEntity<ResponseDTO> findHeartDetail(@PathVariable("heartId") Long heartId, HttpServletRequest httpServletRequest) {
         User user = (User) httpServletRequest.getAttribute("user");
-        Data data = heartService.findHeartDetail(user, heartId);
+        Data data = heartFacade.findHeartDetail(user, heartId);
         ResponseDTO responseDTO = ResponseDTO.builder().status(MESSAGE_SUCCESS).message(FIND_HEART_DETAIL_SUCCESS).data(data).build();
         return new ResponseEntity<>(responseDTO, HttpStatus.OK);
     }
@@ -58,7 +57,7 @@ public class HeartController {
     @PostMapping("/user-hearts/{heartId}")
     public ResponseEntity<ResponseDTO> saveUserHearts(@PathVariable("heartId") Long heartId, HttpServletRequest httpServletRequest) {
         User user = (User) httpServletRequest.getAttribute("user");
-        userHeartService.saveUserHearts(user.getId(), heartId);
+        heartFacade.saveUserHearts(user.getId(), heartId);
         ResponseDTO responseDTO = ResponseDTO.builder().status(MESSAGE_SUCCESS).message(SAVE_USER_HEART_SUCCESS).build();
         return new ResponseEntity<>(responseDTO, HttpStatus.OK);
     }
