+++ backend/src/main/java/com/chillin/hearting/api/service/UserService.java
@@ -286,5 +286,13 @@ public class UserService {
         CookieUtil.deleteCookie(httpServletRequest, httpServletResponse, REFRESH_TOKEN);
     }
 
+    @Transactional(readOnly = true)
+    public User findById(String userId) {
+        return userRepository.findById(userId).orElseThrow(UserNotFoundException::new);
+    }
 
+    @Transactional
+    public User save(User user) {
+        return userRepository.save(user);
+    }
 }
